+++ src/main/java/roomescape/config/ReservationServiceConfig.java
@@ -3,7 +3,7 @@ package roomescape.config;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import roomescape.repository.ReservationRepository;
+import roomescape.repository.reservation.ReservationRepository;
 import roomescape.service.ReservationService;
 
 @Configuration
