+++ src/main/java/roomescape/config/ReservationTimeServiceConfig.java
@@ -3,7 +3,7 @@ package roomescape.config;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import roomescape.repository.ReservationTimeRepository;
+import roomescape.repository.reservationtime.ReservationTimeRepository;
 import roomescape.service.ReservationTimeService;
 
 @Configuration
