+++ src/main/java/roomescape/repository/reservation/ReservationMemoryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.repository.reservation;
 
 import java.util.ArrayList;
 import java.util.List;
