+++ src/main/java/roomescape/repository/reservationtime/ReservationTimeMemoryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.repository.reservationtime;
 
 import java.util.ArrayList;
 import java.util.List;
